requirements.txt
# LangChain 核心
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langgraph>=0.1.0
langchain-deepseek>=0.1.0

# 向量数据库
lancedb>=0.5.0
pyarrow>=14.0.0

# Embedding 模型
torch>=2.0.0
sentence-transformers>=2.2.0

# 文档处理
PyPDF2>=3.0.0

# Web 框架
fastapi>=0.104.0
uvicorn>=0.24.0

# 配置管理
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 工具库
requests>=2.30.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
pandas>=1.5.0

# 监控
prometheus-client>=0.17.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0